marcos.cargo: add path parameter in cargo_install (#3)
The `%{cargo_install}` macro is expanded adding a `--path .` paramenter
at the end.  With virtual workspaces we need to specify a different
`path` per binary crate, but this cannot be done if there is already one
`path` parameter.

This patch adds a path parameter (-p) in the cargo_install macro, that
if it is not present will default to the current "--path ." one.  Now an
user can specify a different path parameter:

  %{cargo_install -p binary-crate} --no-default-features

Related with:

  https://github.com/rust-lang/cargo/issues/7599

Signed-off-by: Alberto Planas <aplanas@suse.com>