Always Run Cargo In Release Mode
This commit changes the build flags of the cargo_test RPM macro, Cargo
runs in release mode. This way, the program doesn't get rebuilt between
invocations of cargo. Release is the default target for the cargo
install and listing it would result in an error.  Furthermore, it adds
the option offline to the shared cargo flags, because it's a shared flag
between the macros.